frozen/Adafruit_CircuitPython_UC8151D/requirements.txt
# SPDX-FileCopyrightText: 2022 Alec Delaney, for Adafruit Industries
#
# SPDX-License-Identifier: Unlicense

Adafruit-Blinka
adafruit-blinka-displayio
